Moved utilities to a separate package
