+++ src/main/java/net/hydromatic/sqllogictest/ExecutionOptions.java
@@ -24,6 +24,7 @@ package net.hydromatic.sqllogictest;
 
 import net.hydromatic.sqllogictest.executors.SqlSltTestExecutor;
 
+import net.hydromatic.sqllogictest.util.Utilities;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.BufferedReader;
