+++ src/main/java/net/hydromatic/sqllogictest/executors/ISqlTestOperation.java
@@ -22,7 +22,7 @@
  */
 package net.hydromatic.sqllogictest.executors;
 
-import net.hydromatic.sqllogictest.ICastable;
+import net.hydromatic.sqllogictest.util.ICastable;
 
 /**
  * Base interface for SqlLogicTest operations: either statements or queries.
