+++ src/main/java/net/hydromatic/sqllogictest/util/Utilities.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.hydromatic.sqllogictest;
+package net.hydromatic.sqllogictest.util;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
